♻️ refactor(blazwitcher-extension): simplify search action handling and extract utilities
- Extract buildSearchActionItems and normalizeSearchValue to dedicated util files
- Simplify search shortcuts (Enter / Shift+Enter) and separate into own settings group
- Remove searchValue prop from List and useKeyboardListen, treat SearchAction as regular list items
